third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/tentative/cross-document-traversal-same-document-traversal-expected.txt

This is a testharness.js-based test.
FAIL traversals in the same (back) direction: the second is ignored assert_equals: first load event must be going back (hash) expected "#2" but got ""
FAIL traversals in the same (forward) direction: the second is ignored assert_equals: first load event must be going forward (hash) expected "" but got "#3"
Harness: the test ran to completion.
